Fixed creating table and added raw loading data to table.

+++ src/main/scala/com/eleks/nataliya/bartkiv/Data.scala
@@ -5,5 +5,5 @@ import java.sql.Timestamp
 case class Data(id : Int,
                 value : Double,
                 datetime : Timestamp,
-                latitude : BigDecimal,
-                longitude : BigDecimal)
+                latitude : Double,
+                longitude : Double)
